+++ eol_scons/tools/qt6.py
@@ -3,11 +3,10 @@
 # This source code is licensed under the MIT license found in the LICENSE
 # file in the root directory of this source tree.
 """
-This tool adds Qt6 include paths and libraries to the build
-environment.  Qt6 is similar to Qt4 in that it is divided into many
-different modules, and the modules can be applied to the environment
-individually using either the EnableQtModules() method or by listing the
-module as a tool.  For example, these are equivalent:
+This tool adds Qt6 include paths and libraries to the build environment.  Qt6
+is divided into many different modules, and the modules can be applied to the
+environment individually using either the EnableQtModules() method or by
+listing the module as a tool.  For example, these are equivalent:
 
     qtmods = ['QtSvg', 'QtCore', 'QtGui', 'QtNetwork', 'QtSql', 'QtOpenGL']
     env.EnableQtModules(qtmods)
@@ -35,9 +34,8 @@ then a particular module may not be installed and pkg-config for it will
 fail.
 
 The qt6 tool must be included first to force all the subsequent qt modules
-to be applied as qt6 modules instead of qt4.  The biggest difference is the
-location of the header files and the version-qualified library names like
-libQt6<Module>.
+to be applied as qt6 modules.  The biggest difference is the location of the
+header files and the version-qualified library names like libQt6<Module>.
 """
 
 # Notes on install locations for each environment
@@ -592,9 +590,9 @@ def enable_modules(env, modules, debug=False):
     main entry point enforces a few things before calling the
     platform-specific code:
 
-    The module name must be a Qt module name that is not qualified by the
-    Qt version.  So QtCore is the module name in Qt4, Qt5 and Qt6.  This
-    function specifically rejects module names starting with Qt4, Qt5 or Qt6.
+    The module name must be a Qt module name that is not qualified by the Qt
+    version, so this function specifically rejects module names starting with
+    Qt6.
 
     QT6DIR must be set in the Environment.  If not, then the Qt6 setup in
     generate() above did not succeed, and therefore no Qt6 modules can be
@@ -609,7 +607,7 @@ def enable_modules(env, modules, debug=False):
 
     onefailed = False
     for module in modules:
-        if module.startswith('Qt6') or module.startswith('Qt5'):
+        if module.startswith('Qt6'):
             raise SCons.Errors.StopError(
                 "Qt module names should not be qualified with "
                 "the version: %s" % (module))